Move PeerSet initialization into a submodule.
